Refactor: Allow wildcard in ALLOWED_CLIENT_HOSTS and update dependencies
This commit updates the middleware to allow the use of a wildcard in the ALLOWED_CLIENT_HOSTS setting. Additionally, the dependencies have been updated to require Django version 4.1.4 and support for Python 3.10 has been added. The version of the package in setup.py has also been modified to reflect these changes.